split navigator module and whetstone-runtime module
