+++ build.gradle
@@ -35,6 +35,11 @@ dependencyAnalysis {
             onAny {
                 severity('fail')
             }
+
+            onIncorrectConfiguration {
+                // https://github.com/autonomousapps/dependency-analysis-android-gradle-plugin/issues/522
+                exclude("androidx.compose.runtime:runtime")
+            }
         }
     }
 }
