--- navigator/build.gradle
@@ -1,50 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'org.jetbrains.kotlin.android'
-apply plugin: 'com.vanniktech.maven.publish'
-
-android {
-    compileSdkVersion 31
-
-    defaultConfig {
-        minSdkVersion 21
-    }
-
-    // still needed for Android projects despite toolchain
-    compileOptions {
-        sourceCompatibility(JavaVersion.VERSION_11)
-        targetCompatibility(JavaVersion.VERSION_11)
-    }
-}
-
-// workaround for https://youtrack.jetbrains.com/issue/KT-37652
-android.kotlinOptions.freeCompilerArgs += "-Xexplicit-api=strict"
-
-kotlin {
-    explicitApi()
-
-    jvmToolchain {
-        languageVersion.set(JavaLanguageVersion.of(11))
-    }
-
-    sourceSets.all {
-        languageSettings {
-            optIn("kotlin.RequiresOptIn")
-        }
-    }
-}
-
-// workaround for https://issuetracker.google.com/issues/194113162
-tasks.withType(JavaCompile).configureEach {
-    javaCompiler = javaToolchains.compilerFor {
-        languageVersion = JavaLanguageVersion.of(11)
-    }
-}
-
-dependencies {
-    api "androidx.activity:activity:1.4.0"
-    api "androidx.navigation:navigation-runtime:2.3.5"
-    api "org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.5.2"
-
-    // dependencies to be added by consumers depending on what they use
-    compileOnly "androidx.fragment:fragment:1.3.6"
-}
