+++ navigator/common-compose/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=navigator-runtime-compose
+POM_NAME=Navigator Runtime Compose
+POM_DESCRIPTION=Navigator Runtime Compose
