+++ navigator/common-fragment/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=navigator-runtime-fragment
+POM_NAME=Navigator Runtime Fragment
+POM_DESCRIPTION=Navigator Runtime Fragment
