+++ navigator/common/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=navigator-common
+POM_NAME=Navigator Common
+POM_DESCRIPTION=Navigator Common
