--- navigator/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=navigator
-POM_NAME=Navigator
-POM_DESCRIPTION=Navigator
