+++ whetstone/runtime-compose/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=whetstone-runtime-compose
+POM_NAME=Whetstone Runtime Compose
+POM_DESCRIPTION=Whetstone Runtime Compose
