+++ whetstone/runtime-fragment/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=whetstone-runtime-fragment
+POM_NAME=Whetstone Runtime Fragment
+POM_DESCRIPTION=Whetstone Runtime Fragment
