+++ whetstone/runtime/src/main/java/com/freeletics/mad/whetstone/internal/ViewModelProvider.kt
@@ -5,7 +5,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalSavedStateRegistryOwner
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.AbstractSavedStateViewModelFactory
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
@@ -13,28 +12,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.LocalViewModelStoreOwner
 import androidx.navigation.NavBackStackEntry
 import androidx.savedstate.SavedStateRegistryOwner
-import javax.inject.Qualifier
 import kotlin.reflect.KClass
-
-/**
- * Creates a [ViewModelProvider] for the given [fragment] that uses a custom
- * [AbstractSavedStateViewModelFactory]. That factory will use [scope] to lookup the required
- * dependencies interface [D] implementation from either the `Activity` or the `Application`
- * `Context`. The [D] instance will then, together with the [SavedStateHandle] be passed to
- * [factory] to instantiate the [ViewModel].
- *
- * To be used in generated code.
- */
-@InternalWhetstoneApi
-public fun <D> viewModelProvider(
-    fragment: Fragment,
-    scope: KClass<*>,
-    factory: (D, SavedStateHandle) -> ViewModel
-): ViewModelProvider {
-    val context = fragment.requireContext()
-    val viewModelFactory = SimpleSavedStateViewModelFactory(fragment, context, scope, factory)
-    return ViewModelProvider(fragment, viewModelFactory)
-}
 
 /**
  * Creates a [ViewModelProvider] for the given [fragment] that uses a custom
@@ -80,7 +58,8 @@ public fun <C> viewModelProvider(
     return ViewModelProvider(entry, viewModelFactory)
 }
 
-private class SimpleSavedStateViewModelFactory<D>(
+@InternalWhetstoneApi
+public class SimpleSavedStateViewModelFactory<D>(
     savedStateRegistryOwner: SavedStateRegistryOwner,
     private val context: Context,
     private val scope: KClass<*>,
